requirements.txt
altair
annotated-types
anyio
appnope
asttokens
attrs
azure-common
azure-core
azure-search-documents
azure-storage-blob
blinker
cachetools
certifi
cffi
charset-normalizer
click
comm
cryptography
debugpy
decorator
distro
executing
gitdb
gitpython
h11
httpcore
httpx
idna
ipykernel
ipython
ipython-pygments-lexers
isodate
jedi
jinja2
jiter
jsonschema
jsonschema-specifications
jupyter-client
jupyter-core
markupsafe
matplotlib-inline
narwhals
nest-asyncio
numpy
openai
packaging
pandas
parso
pexpect
pillow
pip
platformdirs
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pycparser
pydantic
pydantic-core
pydeck
pygments
python-dateutil
python-dotenv
pytz
pyzmq
referencing
requests
rpds-py
six
smmap
sniffio
stack-data
streamlit
tenacity
toml
tornado
tqdm
traitlets
typing-extensions
typing-inspection
tzdata
urllib3
wcwidth
dotenv